+++ src/main/java/dev/willyelton/crystal_tools/common/levelable/armor/LevelableArmor.java
@@ -200,4 +200,10 @@ public class LevelableArmor extends ArmorItem implements LevelableItem, Equipabl
     public boolean isBookEnchantable(ItemStack stack, ItemStack book) {
         return CrystalToolsConfig.ENCHANT_TOOLS.get();
     }
+
+    @Override
+    public boolean shouldCauseBlockBreakReset(ItemStack oldStack, ItemStack newStack) {
+        // Just ignore data components for now
+        return !newStack.is(oldStack.getItem());
+    }
 }