+++ changelog.md
@@ -1,4 +1,11 @@
 # Crystal Tools Change Log
+## 2.2.3
+### Additions
+- Adds a new keybind (`R`). This will automatically use a Crystal Rocket from your inventory while you are flying.
+
+### Fixes
+- Fixes issue with reequip animations playing too often when the repair in hand config option was true
+
 ## 2.2.2
 ### Additions
 - Adds in-game documentation using [GuideMe](https://www.curseforge.com/minecraft/mc-mods/guideme)